avoid a printf in the library
In general, we don't want libraries to send data to the standard file
descriptors.  There are more that need fixing.

Note: this introduces a new API (psl_suffix_count() and
psl_suffix_exception_count) to enable the same sort of output from the
test.  But this new API seems to imply the internal structure of the
public suffix list.

Do we want to expose this API?  There could be some other PSL
mechanism (e.g. DBOUND) that doesn't have these counts, and a drop-in
replacement would not know what to return here.