fix debian/rules clean
For some reason i don't understand, make -j4 distclean (which is
invoked by the clean target of debian/rules) fails because the
"missing" file isn't present in the source root.

this avoids that problem, using the version of missing/ found in
libtool (which is the same version in automake-1.15 as of this
writing).