meson: mark all test-related build artifacts as not built by default
They will not be top-level dependencies of `ninja all`, but still
participate in the dependency graph and be built as a prerequisite for
any target that requires it -- for example, the test target.

Due to meson inefficiencies, all test prerequisites are *also* built as
part of `all`, but this is probably going to change in the future. For
now, this does no harm, and means that in the future, when Meson changes
test prerequisites to only be built by `ninja test` and not `ninja all`,
libpsl will automatically start to skip building the tests by default,
even when tests are enabled.